+++ storybook/src/components/form/form-group-textarea.css-template.ts
@@ -44,7 +44,7 @@ export const cssFormGroupTextarea: ComponentImplementation<FormGroupTextarea<Tem
               rows=${ifDefined(formGroup.rows)}
               aria-describedby=${ifDefined(ariaDescribedBy)}
               aria-errormessage=${ifDefined(ariaErrorMessage)}
-              aria-invalid=${ifDefined(formGroup.state)}
+              aria-invalid=${formGroup.state === "invalid"}
               ?disabled=${formGroup.disabled}
               ?readonly=${formGroup.readonly}
               ?required=${formGroup.required}